MTP now run correctly on wasi
