+++ src/Platform/Microsoft.Testing.Platform/OutputDevice/OutputDeviceManager.cs
@@ -2,6 +2,9 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using Microsoft.Testing.Platform.Logging;
+#if !NET7_0_OR_GREATER
+using Microsoft.Testing.Platform.Resources;
+#endif
 using Microsoft.Testing.Platform.ServerMode;
 using Microsoft.Testing.Platform.Services;
 
@@ -41,27 +44,49 @@ internal sealed class PlatformOutputDeviceManager
     }
 
     public static IPlatformOutputDevice GetDefaultTerminalOutputDevice(ServiceProvider serviceProvider)
-        => RuntimeInformation.IsOSPlatform(OSPlatform.Create("BROWSER")) || RuntimeInformation.IsOSPlatform(OSPlatform.Create("WASI"))
-            ? new BrowserOutputDevice(
+    {
+        if (OperatingSystem.IsBrowser())
+        {
+#if NET7_0_OR_GREATER
+            return new BrowserOutputDevice(
                 serviceProvider.GetConsole(),
                 serviceProvider.GetTestApplicationModuleInfo(),
                 serviceProvider.GetAsyncMonitorFactory().Create(),
                 serviceProvider.GetRuntimeFeature(),
                 serviceProvider.GetEnvironment(),
                 serviceProvider.GetPlatformInformation(),
-                serviceProvider.GetRequiredService<IStopPoliciesService>())
-            : new TerminalOutputDevice(
+                serviceProvider.GetRequiredService<IStopPoliciesService>());
+#else
+            throw new PlatformNotSupportedException(PlatformResources.BrowserPlatformNotSupportedOnOlderFrameworks);
+#endif
+        }
+
+        if (OperatingSystem.IsWasi())
+        {
+            return new WasiOutputDevice(
                 serviceProvider.GetConsole(),
                 serviceProvider.GetTestApplicationModuleInfo(),
-                serviceProvider.GetTestHostControllerInfo(),
                 serviceProvider.GetAsyncMonitorFactory().Create(),
                 serviceProvider.GetRuntimeFeature(),
                 serviceProvider.GetEnvironment(),
                 serviceProvider.GetPlatformInformation(),
-                serviceProvider.GetCommandLineOptions(),
-                serviceProvider.GetFileLoggerInformation(),
-                serviceProvider.GetLoggerFactory(),
-                serviceProvider.GetClock(),
-                serviceProvider.GetRequiredService<IStopPoliciesService>(),
-                serviceProvider.GetTestApplicationCancellationTokenSource());
+                serviceProvider.GetRequiredService<IStopPoliciesService>());
+        }
+
+        // Default to terminal output device for other platforms.
+        return new TerminalOutputDevice(
+            serviceProvider.GetConsole(),
+            serviceProvider.GetTestApplicationModuleInfo(),
+            serviceProvider.GetTestHostControllerInfo(),
+            serviceProvider.GetAsyncMonitorFactory().Create(),
+            serviceProvider.GetRuntimeFeature(),
+            serviceProvider.GetEnvironment(),
+            serviceProvider.GetPlatformInformation(),
+            serviceProvider.GetCommandLineOptions(),
+            serviceProvider.GetFileLoggerInformation(),
+            serviceProvider.GetLoggerFactory(),
+            serviceProvider.GetClock(),
+            serviceProvider.GetRequiredService<IStopPoliciesService>(),
+            serviceProvider.GetTestApplicationCancellationTokenSource());
+    }
 }
