+++ src/Platform/Microsoft.Testing.Platform/Resources/PlatformResources.resx
@@ -117,6 +117,9 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
+  <data name="BrowserPlatformNotSupportedOnOlderFrameworks" xml:space="preserve">
+    <value>Browser platform is only supported on .NET 7.0 or greater.</value>
+  </data>
   <data name="CannotCreateTestExecutionFilterErrorMessage" xml:space="preserve">
     <value>Failed to create a test execution filter</value>
   </data>
